Harden security, fix bugs, simplify logging, and add 600+ tests
- Fix CDF boundary, int32 overflow, cache key collision, and weekly fiscal KeyError
- Prevent command injection, path traversal, SQL/TMDL injection, and CORS hardening
- Add thread safety locks, specific exception types, and input size limits
- Simplify budget log output to show filename only
- Restore name_pools default_folder fallback
- Add 5 new test files (dimensions, sales logic, packaging, utils, web routes)

+++ src/facts/inventory/micro_agg.py
@@ -84,5 +84,5 @@ def micro_aggregate_inventory(
         "store_key": out_store.astype(np.int32),
         "year": (out_year_idx + min_year).astype(np.int16),
         "month": (out_month_idx + 1).astype(np.int8),
-        "quantity_sold": sum_qty[indices].astype(np.int32),
+        "quantity_sold": sum_qty[indices].astype(np.int64),
     }
